improvements and bugfixes to TaskScheduler and parfor. parfor now captures loop bounds by copy rather than using the forward_as_tuple during TaskScheduler::enqueue(). This accidentally gave the loop bounds reference semantics, which essentially made only the last loop block be executed.
Also established new pattern of creating a vector of thread_local_variables (see compute_centroids). This
is just a vector of a locally-defined struct which holds objects intended for thread-local use within
a TaskScheduler worker thread. The vector is captured by reference and access is obtained within a task
via the "worker_global::worker_id" variable. This is a thread_local integer (int) that is set to the thread's
index within the TaskScheduler::workers array. This is useful for ElementFactory objects, reduction variables,
etc, which have high copy overhead and protection via mutex would destroy parallelism. Making a copy per task
is prohibitively expensive for many structures, so capturing via copy and copying the lambda for each task
is inefficient.